+++ src/model/activity_map.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use serde::Deserialize;
 use tabled::Tabled;
 
-use crate::deserialize::null_to_default;
+use crate::util::deserialize::null_to_default;
 
 #[derive(Deserialize)]
 pub struct ActivityMaps {
