+++ src/model/device.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 use tabled::Tabled;
 
-use crate::deserialize::null_to_default;
+use crate::util::deserialize::null_to_default;
 
 #[derive(Default, Deserialize, Tabled)]
 #[serde(default)]
