+++ src/util/mod.rs
@@ -1,3 +1,4 @@
+pub mod deserialize;
 pub mod file;
 pub mod print;
 pub mod time;
